refactor(evalhubclient): move package from internal/evalhub_client to pkg/evalhubclient
Relocates the EvalHub REST API client to pkg/ so it can be imported by
external consumers (e.g. the evalhub-mcp server) without going through
the internal/ visibility restriction. Package name updated from the
underscore form evalhub_client to the idiomatic evalhubclient.
