+++ pkg/evalhubclient/client_test.go
@@ -1,4 +1,4 @@
-package evalhub_client
+package evalhubclient
 
 import (
 	"context"
